move exception mappers to different package to avoid registering them when scanning com.nitorcreations.rest.config package
