+++ nflow-jetty/src/main/java/com/nitorcreations/nflow/jetty/config/NflowJettyConfiguration.java
@@ -34,11 +34,11 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider;
 import com.nitorcreations.nflow.engine.internal.config.NFlow;
-import com.nitorcreations.nflow.jetty.validation.CustomValidationExceptionMapper;
-import com.nitorcreations.nflow.rest.config.BadRequestExceptionMapper;
+import com.nitorcreations.nflow.jetty.mapper.BadRequestExceptionMapper;
+import com.nitorcreations.nflow.jetty.mapper.CustomValidationExceptionMapper;
+import com.nitorcreations.nflow.jetty.mapper.NotFoundExceptionMapper;
 import com.nitorcreations.nflow.rest.config.CorsHeaderContainerResponseFilter;
 import com.nitorcreations.nflow.rest.config.DateTimeParamConverterProvider;
-import com.nitorcreations.nflow.rest.config.NotFoundExceptionMapper;
 import com.nitorcreations.nflow.rest.config.RestConfiguration;
 import com.nitorcreations.nflow.rest.v1.ArchiveResource;
 import com.nitorcreations.nflow.rest.v1.StatisticsResource;
